The API gateway implementation in progress. The two mirco services will also be implemented soon. The DB will be developed using No SQL mongoDB as per request and attached to one of the services. The other one will work with the Slack API endpoints.
